+++ com.unity.render-pipelines.high-definition/CHANGELOG.md
@@ -51,6 +51,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Fixed emission node breaking the main shader graph preview in certain conditions.
 - Fixed checkout of baked probe asset when baking probes.
 - Fixed invalid gizmo position for rotated ReflectionProbe
+- Fixed multi-edition of material's SurfaceType and RenderingPath
 
 ### Changed
 - DensityVolume scripting API will no longuer allow to change between advance and normal edition mode